requirements.txt
python-telegram-bot
ollama
deep-translator
pathlib